PR feedback 반영
