+++ Assets/Tests/PlayMode/MainStageManagerTest.cs
@@ -13,8 +13,6 @@ public class MainStageManagerTest
     private MainStageManager mainStageManager;
     private GameObject gameManagerObject;
     private GameManager gameManager;
-    private GameObject playerGameObject;
-    private MainStagePlayer mainStagePlayer;
 
     [SetUp]
     public void Setup()
@@ -26,8 +24,6 @@ public class MainStageManagerTest
 
         gameManager.LoadMainStage();
         mainStageManager = GameObject.FindObjectOfType<MainStageManager>();
-
-        mainStagePlayer = GameObject.FindObjectOfType<MainStagePlayer>();
     }
 
     [TearDown]
@@ -52,35 +48,6 @@ public class MainStageManagerTest
 
         // Assert
         Assert.IsTrue(mainStageManager.IsStageComplete());
-    }
-
-    [UnityTest]
-    public IEnumerator TestLaserFire()
-    {
-        gameManager.LoadMainStage();
-        mainStageManager = GameObject.FindObjectOfType<MainStageManager>();
-        mainStagePlayer = GameObject.FindObjectOfType<MainStagePlayer>();
-
-        // 초기 프로젝타일 카운트 확인
-        int initialProjectileCount = GameObject.FindGameObjectsWithTag("Laser").Length;
-        yield return null;
-        
-        mainStagePlayer = GameObject.FindObjectOfType<MainStagePlayer>();
-        mainStagePlayer.projectileSpawnPoint = mainStagePlayer.transform;
-
-        // 레이저 발사
-        mainStagePlayer.FireLaser();
-        yield return new WaitForSeconds(0.1f);
-
-        // 프로젝타일 증가 확인
-        int newProjectileCount = GameObject.FindGameObjectsWithTag("Laser").Length;
-        Assert.AreEqual(initialProjectileCount + 1, newProjectileCount);
-
-        yield return new WaitForSeconds(1f);
-
-        // 프로젝타일 파괴 확인
-        newProjectileCount = GameObject.FindGameObjectsWithTag("Laser").Length;
-        Assert.AreEqual(initialProjectileCount, newProjectileCount);
     }
 
     [UnityTest]
