+++ Assets/Tests/PlayMode/MainStagePlayerTest.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 882e441f85a99e946908c6fa4e7a7cb7
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
